+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelogs
 
+## v1.4.1
+
 ## v1.4.0
 
 New commands: